backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.7
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
email-validator>=2.0.0
jinja2>=3.1.2
weasyprint>=58.1
psycopg2-binary>=2.9.6
pydantic>=1.10.7
pydantic-settings>=2.0.0
bcrypt>=4.0.1
python-dotenv>=1.0.0